Ajustes al main y clases

+++ src/main/java/com/mycompany/divisax/ConversorDivisas.java
@@ -1,13 +1,9 @@
 package com.mycompany.divisax;
 
 public class ConversorDivisas {
-
     
-    
-    public Double convercionDeMoneda(Double rate, int base){
-        
+    public Double convercionDeMoneda(Double rate, Double base){
          double resultado = rate * base;
-        
         return resultado;
     }
